# frontend/kanban-board-client-app/swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban Board API",
    "description": "Advanced Kanban Board API with search, analytics, performance optimization, and comprehensive features. This API exposes only DTOs for clean separation between internal models and API contracts.",
    "contact": {
      "name": "Kanban Board Team",
      "url": "https://kanbanboard.com",
      "email": "support@kanbanboard.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.kanbanboard.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/attachments": {
      "post": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "createAttachment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "createAuthenticationToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/boards/{boardId}/analytics": {
      "get": {
        "tags": [
          "board-controller"
        ],
        "operationId": "getBoardAnalytics",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BoardAnalytics"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/attachments/{id}": {
      "get": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "getAttachmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "deleteAttachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/attachments/user/{userId}": {
      "get": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "getAttachmentsByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/attachments/task/{taskId}": {
      "get": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "getAttachmentsByTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/attachments/content-type/{contentType}": {
      "get": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "getAttachmentsByContentType",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/user/{userId}": {
      "get": {
        "tags": [
          "analytics-controller"
        ],
        "operationId": "getUserAnalytics",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserAnalytics"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/productivity/{boardId}": {
      "get": {
        "tags": [
          "analytics-controller"
        ],
        "operationId": "getProductivityMetrics",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductivityMetrics"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/my-productivity": {
      "get": {
        "tags": [
          "analytics-controller"
        ],
        "operationId": "getMyProductivityMetrics",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductivityMetrics"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/board/{boardId}": {
      "get": {
        "tags": [
          "analytics-controller"
        ],
        "operationId": "getBoardAnalytics_1",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BoardAnalytics"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RoleDTO": {
        "type": "object",
        "description": "Role data transfer object for API communication",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the role",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the role",
            "example": "ADMIN"
          }
        },
        "required": [
          "name"
        ]
      },
      "UserDTO": {
        "type": "object",
        "description": "User data for registration",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the user",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username for authentication",
            "example": "john_doe",
            "maxLength": 50,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "User password (only for registration/update)",
            "example": "password123",
            "maxLength": 100,
            "minLength": 6
          },
          "displayName": {
            "type": "string",
            "description": "Display name shown to other users",
            "example": "John Doe",
            "maxLength": 100,
            "minLength": 2
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "avatar": {
            "type": "string",
            "description": "URL to user's avatar image",
            "example": "https://example.com/avatar.jpg"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last login",
            "example": "2024-01-15T10:30:00"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when user was created",
            "example": "2024-01-01T00:00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when user was last updated",
            "example": "2024-01-15T10:30:00"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the user account is enabled",
            "example": true
          },
          "name": {
            "type": "string",
            "description": "Legacy name field (deprecated)",
            "example": "John Doe"
          },
          "roles": {
            "type": "array",
            "description": "List of roles assigned to the user",
            "items": {
              "$ref": "#/components/schemas/RoleDTO"
            }
          }
        },
        "required": [
          "displayName",
          "password",
          "username"
        ]
      },
      "AttachmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "filename": {
            "type": "string",
            "minLength": 1
          },
          "originalFilename": {
            "type": "string",
            "minLength": 1
          },
          "fileUrl": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string",
            "minLength": 1
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "contentType",
          "fileSize",
          "fileUrl",
          "filename",
          "originalFilename"
        ]
      },
...
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "BoardAnalytics": {
        "type": "object",
        "properties": {
          "totalTasks": {
            "type": "integer",
            "format": "int32"
          },
          "totalColumns": {
            "type": "integer",
            "format": "int32"
          },
          "totalUsers": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserAnalytics": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "assignedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "createdTasks": {
            "type": "integer",
            "format": "int32"
          },
          "completedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "overdueTasks": {
            "type": "integer",
            "format": "int32"
          },
          "completionRate": {
            "type": "number",
            "format": "double"
          },
          "statusDistribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          },
          "priorityDistribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "ProductivityMetrics": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "integer",
            "format": "int64"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "tasksCreated": {
            "type": "integer",
            "format": "int32"
          },
          "tasksCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "completionRate": {
            "type": "number",
            "format": "double"
          },
          "averageTaskDuration": {
            "type": "number",
            "format": "double"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}